Move externals/test_common.go into its own package
