+++ go/chat/boxer_test.go
@@ -15,7 +15,7 @@ import (
 	"github.com/keybase/client/go/chat/signencrypt"
 	"github.com/keybase/client/go/chat/types"
 	"github.com/keybase/client/go/engine"
-	"github.com/keybase/client/go/externals"
+	"github.com/keybase/client/go/externalstest"
 	"github.com/keybase/client/go/kbtest"
 	"github.com/keybase/client/go/libkb"
 	"github.com/keybase/client/go/protocol/chat1"
@@ -67,7 +67,7 @@ func textMsgWithHeader(t *testing.T, text string, header chat1.MessageClientHead
 }
 
 func setupChatTest(t *testing.T, name string) (*kbtest.ChatTestContext, *Boxer) {
-	tc := externals.SetupTest(t, name, 2)
+	tc := externalstest.SetupTest(t, name, 2)
 
 	// use an insecure triplesec in tests
 	tc.G.NewTriplesec = func(passphrase []byte, salt []byte) (libkb.Triplesec, error) {
