+++ go/chat/storage/common_test.go
@@ -3,13 +3,13 @@ package storage
 import (
 	"testing"
 
-	"github.com/keybase/client/go/externals"
+	"github.com/keybase/client/go/externalstest"
 	"github.com/keybase/client/go/libkb"
 	insecureTriplesec "github.com/keybase/go-triplesec-insecure"
 )
 
 func setupCommonTest(t testing.TB, name string) libkb.TestContext {
-	tc := externals.SetupTest(t, name, 2)
+	tc := externalstest.SetupTest(t, name, 2)
 
 	// use an insecure triplesec in tests
 	tc.G.NewTriplesec = func(passphrase []byte, salt []byte) (libkb.Triplesec, error) {
