+++ go/externalstest/test_common.go
@@ -3,9 +3,10 @@
 
 // +build !production
 
-package externals
+package externalstest
 
 import (
+	"github.com/keybase/client/go/externals"
 	"github.com/keybase/client/go/libkb"
 	"github.com/keybase/client/go/pvlsource"
 	"github.com/keybase/client/go/uidmap"
@@ -14,7 +15,7 @@ import (
 func SetupTest(tb libkb.TestingTB, name string, depth int) (tc libkb.TestContext) {
 	ret := libkb.SetupTest(tb, name, depth+1)
 
-	ret.G.SetServices(GetServices())
+	ret.G.SetServices(externals.GetServices())
 	ret.G.SetUIDMapper(uidmap.NewUIDMap(10000))
 	pvlsource.NewPvlSourceAndInstall(ret.G)
 	return ret
